Fixed errors, everything compiles

+++ Assignment_3/GitGud/GitGud_CS311_assignment3/src/processor/pipeline/EX_MA_LatchType.java
@@ -39,7 +39,7 @@ public class EX_MA_LatchType {
 		this.pc = pc;
 	}
 
-	public int getAluResult() {
+	public long getAluResult() {
 		return this.aluResult;
 	}
