+++ Assignment_3/GitGud/GitGud_CS311_assignment3/src/processor/pipeline/MA_RW_LatchType.java
@@ -52,7 +52,7 @@ public class MA_RW_LatchType {
 	}
 
 
-	public int getAluResult() {
+	public long getAluResult() {
 		return this.aluResult;
 	}
